Split up Location into its own file.
