+++ src/ast2ram/AstToRamTranslator.cpp
@@ -62,7 +62,9 @@
 #include "ast/utility/Utils.h"
 #include "ast/utility/Visitor.h"
 #include "ast2ram/ClauseTranslator.h"
+#include "ast2ram/Location.h"
 #include "ast2ram/ProvenanceClauseTranslator.h"
+#include "ast2ram/ValueIndex.h"
 #include "parser/SrcLocation.h"
 #include "ram/Aggregate.h"
 #include "ram/AutoIncrement.h"
