+++ src/ast2ram/AstToRamTranslator.h
@@ -184,52 +184,4 @@ private:
     Own<ram::Statement> makeNegationSubproofSubroutine(const ast::Clause& clause);
 };
 
-/**
- * Concrete attribute
- */
-struct AstToRamTranslator::Location {
-    int identifier{};
-    int element{};
-    Own<ram::RelationReference> relation{nullptr};
-
-    Location() = default;
-
-    Location(int ident, int elem, Own<ram::RelationReference> rel = nullptr)
-            : identifier(ident), element(elem), relation(std::move(rel)) {}
-
-    Location(const Location& l) : identifier(l.identifier), element(l.element) {
-        if (l.relation != nullptr) {
-            relation = souffle::clone(l.relation);
-        }
-    }
-
-    Location& operator=(Location other) {
-        identifier = other.identifier;
-        element = other.element;
-        relation = std::move(other.relation);
-        return *this;
-    }
-
-    bool operator==(const Location& loc) const {
-        return identifier == loc.identifier && element == loc.element;
-    }
-
-    bool operator!=(const Location& loc) const {
-        return !(*this == loc);
-    }
-
-    bool operator<(const Location& loc) const {
-        return identifier < loc.identifier || (identifier == loc.identifier && element < loc.element);
-    }
-
-    void print(std::ostream& out) const {
-        out << "(" << identifier << "," << element << ")";
-    }
-
-    friend std::ostream& operator<<(std::ostream& out, const Location& loc) {
-        loc.print(out);
-        return out;
-    }
-};
-
 }  // namespace souffle::ast2ram
